requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.101.0
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.43
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
WTForms==3.1.2
itsdangerous==2.2.0